test: Ensure local dependencies are up-to-date
While jest will run a package and its tests without needing a build
step, it will use the last-built version of the other packages for its
dependencies, so we're not actually guaranteed to be testing the latest
state unless we build all of a package's local dependencies before we
start testing it.

The `test:watch` scripts have been removed, as a watch mode that tests
without building is not reliable.

I don't think any local packages are getting injected, but I added
`syncInjectedDepsAfterScripts` to `pnpm-workspace.yaml` to be safe.

Since we're running tests in parallel, individual tests run slower, so
I had to increase some timeouts.